prevent unnecessary re-renders by memoizing Filter and using stable update callback
