+++ src/components/ProductTable/Filters.tsx
@@ -1,21 +1,18 @@
 import { RotateCcw } from "lucide-react"
 
-import { FilterInputState, FilterOption } from "@/lib/types"
+import { FilterOption } from "@/lib/types"
 
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from "@/components/ui/accordion"
+import { Accordion } from "@/components/ui/accordion"
 import { Button } from "@/components/ui/buttons/Button"
+
+import Filter from "./Filter"
 
 import { useTranslation } from "@/hooks/useTranslation"
 
 interface PresetFiltersProps {
   filters: FilterOption[]
   activeFiltersCount: number
-  setFilters: (filterOptions: FilterOption[]) => void
+  setFilters: (filter: FilterOption, filterIndex: number) => void
   resetFilters: () => void
 }
 
@@ -26,53 +23,6 @@ const Filters = ({
   activeFiltersCount,
 }: PresetFiltersProps) => {
   const { t } = useTranslation("table")
-
-  const updateFilterState = (
-    filterIndex: number,
-    itemIndex: number,
-    newInputState: FilterInputState,
-    optionIndex?: number
-  ) => {
-    const updatedFilters = filters.map((filter, idx) => {
-      if (idx !== filterIndex) return filter
-
-      const updatedItems = filter.items.map((item, i) => {
-        if (i === itemIndex) {
-          if (typeof optionIndex !== "undefined") {
-            const updatedOptions = item.options.map((option, j) => {
-              if (j === optionIndex) {
-                return {
-                  ...option,
-                  inputState: newInputState,
-                }
-              }
-              return option
-            })
-            return {
-              ...item,
-              options: updatedOptions,
-            }
-          }
-          return {
-            ...item,
-            inputState: newInputState,
-            options: item.options.map((option) => {
-              return {
-                ...option,
-                inputState: newInputState,
-              }
-            }),
-          }
-        }
-        return item
-      })
-      return {
-        ...filter,
-        items: updatedItems,
-      }
-    })
-    setFilters(updatedFilters)
-  }
 
   return (
     <div className="w-full lg:w-80" data-testid="filters-container">
@@ -95,51 +45,14 @@ const Filters = ({
         defaultValue={filters.map((_, idx) => `item ${idx}`)}
       >
         {filters.map((filter, filterIndex) => {
-          if (filter.showFilterOption) {
-            return (
-              <AccordionItem
-                key={filterIndex}
-                value={`item ${filterIndex}`}
-                className="bg-background-highlight p-6"
-              >
-                <AccordionTrigger className="border-b md:px-0">
-                  <p className="text-base font-bold text-body">
-                    {filter.title}
-                  </p>
-                </AccordionTrigger>
-                <AccordionContent className="p-0 md:p-0">
-                  {filter.items.map((item, itemIndex) => {
-                    return (
-                      <div
-                        key={itemIndex}
-                        className={`${item.options.length ? "pb-0" : "pb-4"}`}
-                      >
-                        {item.input(
-                          filterIndex,
-                          itemIndex,
-                          item.inputState,
-                          updateFilterState
-                        )}
-                        {item.inputState === true && item.options.length ? (
-                          <div className="flex flex-row gap-6 px-2 pb-4">
-                            {item.options.map((option, optionIndex) => {
-                              return option.input(
-                                filterIndex,
-                                itemIndex,
-                                optionIndex,
-                                option.inputState,
-                                updateFilterState
-                              )
-                            })}
-                          </div>
-                        ) : null}
-                      </div>
-                    )
-                  })}
-                </AccordionContent>
-              </AccordionItem>
-            )
-          }
+          return (
+            <Filter
+              key={filterIndex}
+              filter={filter}
+              filterIndex={filterIndex}
+              onChange={setFilters}
+            />
+          )
         })}
       </Accordion>
     </div>
