+++ src/components/ProductTable/index.tsx
@@ -41,6 +41,40 @@ interface ProductTableProps<T extends { id: string }> {
   matomoEventCategory: string
 }
 
+const getActiveFiltersCount = (filters: FilterOption[]) => {
+  return filters.reduce((count, filter) => {
+    return (
+      count +
+      filter.items.reduce((itemCount, item) => {
+        if (item.options && item.options.length > 0) {
+          return (
+            itemCount +
+            item.options.filter(
+              (option) =>
+                typeof option.inputState === "boolean" && option.inputState
+            ).length
+          )
+        }
+        if (Array.isArray(item.inputState) && item.inputState.length > 0) {
+          return itemCount + 1
+        }
+
+        if (
+          typeof item.inputState === "string" &&
+          item.filterKey !== "languages"
+        ) {
+          return itemCount + 1
+        }
+
+        return (
+          itemCount +
+          (typeof item.inputState === "boolean" && item.inputState ? 1 : 0)
+        )
+      }, 0)
+    )
+  }, 0)
+}
+
 const ProductTable = <T extends { id: string }>({
   data,
   filters: initialFilters,
@@ -200,7 +234,14 @@ const ProductTable = <T extends { id: string }>({
 
   // Update activePresets based on current filters
   const updateFilters = useCallback(
-    (filters: FilterOption[]) => {
+    (filter: FilterOption, filterIndex: number) => {
+      setFilters((prevFilters) => {
+        return prevFilters.map((prevFilter, idx) => {
+          if (idx !== filterIndex) return prevFilter
+          return filter
+        })
+      })
+
       const currentFilters = {}
 
       filters.forEach((filter) => {
@@ -237,8 +278,6 @@ const ProductTable = <T extends { id: string }>({
         []
       )
 
-      setFilters(filters)
-
       setActivePresets((prevActivePresets) => {
         const newActivePresets = [
           ...new Set([...prevActivePresets, ...presetsToApply]),
@@ -246,43 +285,8 @@ const ProductTable = <T extends { id: string }>({
         return newActivePresets.filter((idx) => presetsToApply.includes(idx))
       })
     },
-    [presetFilters, setFilters, setActivePresets]
+    []
   )
-
-  // Count active filters
-  const activeFiltersCount = useMemo(() => {
-    return filters.reduce((count, filter) => {
-      return (
-        count +
-        filter.items.reduce((itemCount, item) => {
-          if (item.options && item.options.length > 0) {
-            return (
-              itemCount +
-              item.options.filter(
-                (option) =>
-                  typeof option.inputState === "boolean" && option.inputState
-              ).length
-            )
-          }
-          if (Array.isArray(item.inputState) && item.inputState.length > 0) {
-            return itemCount + 1
-          }
-
-          if (
-            typeof item.inputState === "string" &&
-            item.filterKey !== "languages"
-          ) {
-            return itemCount + 1
-          }
-
-          return (
-            itemCount +
-            (typeof item.inputState === "boolean" && item.inputState ? 1 : 0)
-          )
-        }, 0)
-      )
-    }, 0)
-  }, [filters])
 
   const presetFiltersCounts = useMemo(() => {
     return presetFilters.map((persona) => {
@@ -333,6 +337,11 @@ const ProductTable = <T extends { id: string }>({
       }
     },
     [matomoEventCategory]
+  )
+
+  const activeFiltersCount = useMemo(
+    () => getActiveFiltersCount(filters),
+    [filters]
   )
 
   return (
